pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "writ"
version = "1.0.0"
description = "Transparent AI knowledge graphs from regulatory documents"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "neo4j>=5.0.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pypdf>=4.0.0",
]

[project.optional-dependencies]
langchain = [
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-neo4j>=0.1.0",
    "langchain-community>=0.3.0",
]
gpu = [
    "sentence-transformers>=2.2.0",
    "torch>=2.0.0",
]
llamaparse = [
    "llama-parse>=0.4.0",
]
dev = [
    "pytest>=7.0.0",
    "boto3>=1.28.0",
]
all = [
    "writ[langchain,gpu,llamaparse,dev]",
]

[tool.setuptools.packages.find]
include = ["lib*", "query*", "utils*", "validation*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "unit: fast tests with no external dependencies",
    "integration: requires running Neo4j (docker-compose up)",
]
